requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
algo-tools @ git+http://10.0.3.101/kaiwu/agent/algo_tools.git@1fb5908372a731e0f941c84c4373ed5c29bab0b7
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.1
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==21.2.0
bcrypt==3.2.0
beniget==0.4.1
blinker==1.4
bloom==0.12.0
breezy==3.2.1
Brlapi==0.8.3
Brotli==1.0.9
builtin-interfaces==1.2.1
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.2
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
composition-interfaces==1.2.1
configobj==5.0.6
cov-core==1.15.0
coverage==6.2
cpuset==1.6
cryptography==3.4.8
cupshelpers==1.0
cv-bridge==3.2.1
cycler==0.11.0
dataclasses-json==0.6.7
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
demo-nodes-py==0.20.5
Deprecated==1.2.13
diagnostic-msgs==4.8.0
diagnostic-updater==4.0.6
dill==0.4.0
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.10.0
dulwich==0.20.31
duplicity==0.8.21
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.3.0
fastapi==0.115.12
fastbencode==0.0.5
fasteners==0.14.1
fastimport==0.9.14
flake8==4.0.1
fonttools==4.29.1
frozenlist==1.6.0
fs==2.4.12
future==0.18.2
gast==0.5.2
geojson==2.5.0
geometry-msgs==4.8.0
gevent==25.5.1
google-search-results==2.4.2
gpg==1.16.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
icmplib==3.0.4
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
iniconfig==1.1.1
interactive-markers==2.3.2
jeepney==0.7.1
Jetson.GPIO==2.1.7
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==23.5.0
kiwisolver==1.3.2
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-mcp-adapters==0.0.11
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langgraph==0.2.76
langgraph-checkpoint==2.0.26
langgraph-sdk==0.1.70
langsmith==0.3.43
language-selector==0.1
lark==1.1.1
laser-geometry==2.4.0
launch==1.0.9
launch-ros==0.19.9
launch-testing==1.0.9
launch-testing-ros==0.19.9
launch-xml==1.0.9
launch-yaml==1.0.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.1
lockfile==0.12.2
logging-demo==0.20.5
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
map-msgs==2.1.0
MarkupSafe==2.0.1
marshmallow==3.26.1
matplotlib==3.5.1
mccabe==0.6.1
mcp==1.9.1
mercurial==6.1.1
message-filters==4.3.7
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.18
mypy_extensions==1.1.0
nav-msgs==4.8.0
netifaces==0.11.0
nose2==0.9.2
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
opencv-python==4.11.0.86
orjson==3.10.18
ormsgpack==1.10.0
osrf-pycommon==2.1.6
packaging==21.3
pandas==1.3.5
paramiko==2.9.3
patiencediff==0.2.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
propcache==0.3.1
protobuf==5.29.5
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
PyAudio==0.2.14
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pydantic==2.12.3
pydantic-settings==2.9.1
pydantic_core==2.41.4
pydocstyle==6.1.1
pydot==1.4.2
pydub==0.25.1
pyflakes==2.4.0
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyowm==3.3.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-gitlab==2.10.1
python-multipart==0.0.20
python-qt-binding==1.1.2
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.4.0
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
resource-retriever==3.1.3
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rospkg-modules==1.6.0
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scikit-learn==1.6.1
scipy==1.15.3
SecretStorage==3.3.1
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
setuptools-scm==8.3.1
shape-msgs==4.8.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.41
sros2==0.10.6
sse-starlette==2.3.5
starlette==0.46.2
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
supervision==0.25.1
sympy==1.9
systemd-python==234
teleop-twist-keyboard==2.4.0
tenacity==9.1.2
tensorrt==10.3.0
tensorrt_dispatch==10.3.0
tensorrt_lean==10.3.0
tf2-geometry-msgs==0.25.13
tf2-kdl==0.25.13
tf2-msgs==0.25.13
tf2-py==0.25.13
tf2-ros-py==0.25.13
tf2-sensor-msgs==0.25.13
tf2-tools==0.25.13
threadpoolctl==3.6.0
tiktoken==0.9.0
toml==0.10.2
tomli==2.2.1
topic-monitor==0.20.5
torchvision==0.20.1
trajectory-msgs==4.8.0
turtlesim==1.4.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urllib3==1.26.5
urwid==2.1.2
uvicorn==0.34.2
vcstool==0.3.0
vcstools==0.1.42
visualization-msgs==4.8.0
wadllib==1.3.6
websocket==0.2.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.13.3
ws4py==0.6.0
xdg==5
xkit==0.0.0
yarl==1.20.0
zerorpc==0.6.3
zipp==1.0.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
